% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_one_crypto.R
\name{crypto_show_one_ticker}
\alias{crypto_show_one_ticker}
\title{Plot one crypto}
\usage{
crypto_show_one_ticker(ticker_slung, just_show = T)
}
\arguments{
\item{ticker_slung}{The crypto id on coingecko}

\item{just_show}{if just show the plot}
}
\description{
Plot one crypto
}
