% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_crypto_info.R
\docType{data}
\name{data_all_crypto}
\alias{data_all_crypto}
\title{All crypto data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4585 rows and 27 columns.
}
\usage{
data_all_crypto
}
\description{
All crypto data
}
\keyword{datasets}
