% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\name{get_one_stock_data_from_tradingview}
\alias{get_one_stock_data_from_tradingview}
\title{Get one stock info}
\usage{
get_one_stock_data_from_tradingview(x)
}
\arguments{
\item{x}{ticker name}
}
\description{
Get one stock info
}
