% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_plot.R
\name{show_percent_change_of_stocks}
\alias{show_percent_change_of_stocks}
\title{Compare multiple stock prices}
\usage{
show_percent_change_of_stocks(tickers, start_date)
}
\arguments{
\item{tickers}{list of the company tickers}

\item{start_date}{startdate of the data}
}
\description{
Compare multiple stock prices
}
