% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\name{utility_change_col_to_num}
\alias{utility_change_col_to_num}
\title{Changing character column to numeric if it is numeric}
\usage{
utility_change_col_to_num(df)
}
\arguments{
\item{df}{dataframe to check the columns}
}
\description{
Changing character column to numeric if it is numeric
}
